requirements.txt
# binance==0.3
matplotlib==3.3.3
numpy==1.22.2
pandas==1.3.3
pandas_ta==0.3.14b0
python_binance==1.0.12
requests==2.25.1
scipy==1.7.1
tqdm==4.54.1
websocket_client==1.3.2
